refactor: rename pipeline to deals-pipeline for consistency
- Rename sdk/features/pipeline → sdk/features/deals-pipeline
- Rename web/src/components/pipeline → web/src/components/deals-pipeline
- Update imports in page.tsx and pipelineService.ts
- Maintain backward compatibility with deprecated imports
- Keep URL route as /pipeline (user-facing)

This aligns frontend naming with backend naming:
- Backend: backend/features/deals-pipeline/
- Frontend: sdk/features/deals-pipeline/
- API Routes: /api/deals-pipeline/*

Improves developer experience by:
- Consistent naming across frontend and backend
- Easier code navigation and search
- Clear feature identification
- Better onboarding for new developers
