+++ sdk/features/deals-pipeline/components/leads/types.ts
