+++ sdk/features/deals-pipeline/services/pipelineService.ts
