+++ sdk/features/deals-pipeline/store/slices/masterDataSlice.ts
