+++ sdk/features/deals-pipeline/store/slices/pipelineSlice.ts
