+++ sdk/features/deals-pipeline/types.ts
