+++ sdk/features/deals-pipeline/utils/fieldMappings.ts
