+++ sdk/features/deals-pipeline/utils/statusMappings.ts
