+++ web/src/components/deals-pipeline/AddStageDialog.tsx
