+++ web/src/components/deals-pipeline/CreateCardDialog.tsx
