+++ web/src/components/deals-pipeline/EnhancedAddStageDialog.tsx
