+++ web/src/components/deals-pipeline/LeadDetailsDialog.tsx
