+++ web/src/components/deals-pipeline/PipelineBoard.tsx
