+++ web/src/components/deals-pipeline/PipelineBoardHeader.tsx
