+++ web/src/components/deals-pipeline/PipelineBoardSettings.tsx
