+++ web/src/components/deals-pipeline/PipelineFilterDialog.tsx
