+++ web/src/components/deals-pipeline/PipelineKanbanView.tsx
