+++ web/src/components/deals-pipeline/PipelineLeadCard.jsx
