+++ web/src/components/deals-pipeline/PipelineLeadCard.tsx
