+++ web/src/components/deals-pipeline/PipelineListView.tsx
