+++ web/src/components/deals-pipeline/PipelineStageColumn.tsx
