+++ web/src/components/deals-pipeline/SlotBasedPipelineBoard.tsx
