+++ web/src/components/deals-pipeline/config/pipelineConfig.ts
