+++ web/src/components/deals-pipeline/index.ts
