+++ web/src/components/deals-pipeline/slots/CounsellorScheduleSlot.tsx
