+++ web/src/components/deals-pipeline/slots/EducationStudentSlot.tsx
