+++ web/src/components/deals-pipeline/slots/LeadDetailsSlot.tsx
